osc/rdma: ensure local ranks all detect registration failures
If one of the ranks fails to register memory the osc/rdma code will
hang due to a lack of agreement on the state of the window. This
commit updates the shared-memory code to ensure all local ranks have
the save view of the failure/success state of the window.

This commit does not, however, address what happens on failure. Right
now the failure is handed back up to the current error handler.

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>